Change price formatting method
If the price is >1 then we apply the currency format with Intl.NumberFormat and if not we return the price as we got it from the api but concatenated with the fiat currency symbol '$' or '€'. We also cut the decimals in the mktcap and volume values that are added at the end by the Intl.NumberFormat in usd '.00' and eur ',00 €' but in the price we keep them.